// src/app/models/admin.ts
export class Booking {
    EmailID: string;
    NumberOfSeats: number;
    FlightNo: number;
    FlightName: string;
    FlightSource: string;
    FlightDestination: string;
    StartTime: any;
    EndTime: any;
    PassengerDetails : string;
    MealPreference : string;
    PNR : string;
    Status : string;
    ApplyCoupon : string;
}